Optimize GC (#166)
Previously, a GC call would scan the HC table, identifying dead nodes.

Now, whenever a Node is destroyed, if it's ready to be GC'd, it gets added to a list. That list is used at GC time instead of the scan.

This substantially decreases the cost of running GC many times (as the Z# FE does).